+++ test/src/model/reflection.spec.ts
@@ -39,35 +39,35 @@ export function reflectionTests() {
         modelResolverTests();
     });
 }
-        
-function modelResolverTests() {        
+
+function modelResolverTests() {
     describe('modelResolver', () => {
         it('should be possible to obtain the metadata of MyModel', () => {
             var metadata = modelResolver.resolve(MyModel);
-            
+
             expect(metadata.kind).toBe('test::MyModel');
             expect(metadata.superType).toBe(ModelBase);
             expect(metadata.supertypeMeta).toBeNull();
             expect(metadata.ownProperties.keySeq().toArray())
                 .toEqual(['name']);
             expect(metadata.properties.keySeq().toArray())
-                .toEqual(['name']);
+                .toEqual(['id', 'name']);
         });
-        
+
         it('should be possible to obtain the metadata of MySubmodel', () => {
-            var metadata = modelResolver.resolve(MySubmodel);  
+            var metadata = modelResolver.resolve(MySubmodel);
             expect(metadata.kind).toBe('test::MySubmodel');
             expect(metadata.superType).toBe(MyModel);
             expect(metadata.ownProperties.keySeq().toArray())
                 .toEqual(['killTheInnocents']);
             expect(metadata.properties.keySeq().toArray())
-                .toEqual(['name', 'killTheInnocents']);
+                .toEqual(['id', 'name', 'killTheInnocents']);
         });
-        
+
         it('should not be possible to resolve a model with no @Model annotation', () => {
             expect(() => modelResolver.resolve(ModelWithNoMetadata)).toThrow();
         });
-        
+
         it('should not be possible to resolve a model with a >0 argument constructor', () =>{
             expect(() => modelResolver.resolve(ModelWithConstructor)).toThrow();
         });